Created eeg data repository
