+++ mobile/lib/src/application/application.dart
@@ -1,5 +1,5 @@
-export 'bluetooth/bluetoothSelector_cubit.dart';
-export 'bluetooth/bluetoothSelector_state.dart';
+export 'bluetooth/bluetooth_selector_cubit.dart';
+export 'bluetooth/bluetooth_selector_state.dart';
 export 'eeg_data/data_cubit.dart';
 export 'eeg_data/data_states.dart';
 export 'wallets/wallets_cubit.dart';
