+++ mobile/lib/src/application/bluetooth/bluetooth_selector_cubit.dart
@@ -2,7 +2,7 @@ import 'package:bloc/bloc.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_blue/flutter_blue.dart';
 import 'package:polydodo/src/domain/domain.dart';
-import 'bluetoothSelector_state.dart';
+import 'bluetooth_selector_state.dart';
 
 class BluetoothSelectorCubit extends Cubit<BluetoothStates> {
   final IBluetoothRepository _bluetoothRepository;
