+++ mobile/lib/src/domain/domain.dart
@@ -1,5 +1,6 @@
 export 'bluetooth/i_bluetooth_repository.dart';
 export 'eeg_data/eeg_data.dart';
+export 'eeg_data/i_eeg_data_repository.dart';
 export 'wallet/i_wallet_repository.dart';
 export 'wallet/money.dart';
 export 'wallet/owner.dart';
