+++ mobile/lib/src/domain/eeg_data/i_eeg_data_repository.dart
@@ -1,0 +1,8 @@
+abstract class IEEGDataRepository {
+  void createRecordingFromStream(Stream<List<int>> stream);
+  void stopRecordingFromStream();
+
+  // todo: implement export and import
+  void importData();
+  void exportData();
+}
