+++ mobile/lib/src/locator.dart
@@ -2,6 +2,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:get_it/get_it.dart';
 import 'package:polydodo/src/application/eeg_data/data_cubit.dart';
 import 'package:polydodo/src/domain/domain.dart';
+import 'package:polydodo/src/infrastructure/eeg_data_repository.dart';
 import 'package:polydodo/src/infrastructure/infrastructure.dart';
 import 'application/application.dart';
 
@@ -12,6 +13,8 @@ void registerServices() {
   _serviceLocator.registerSingleton<IWalletRepository>(MockWalletRepository());
   _serviceLocator
       .registerSingleton<IBluetoothRepository>(BluetoothRepository());
+  _serviceLocator
+      .registerSingleton<IEEGDataRepository>(EEGDataRepository());
 }
 
 /// This function creates all the BlocProviders used in this app
@@ -29,6 +32,7 @@ List<BlocProvider> createBlocProviders() => [
       BlocProvider<DataCubit>(
         create: (context) => DataCubit(
           _serviceLocator.get<IBluetoothRepository>(),
+          _serviceLocator.get<IEEGDataRepository>(),
         ),
       ),
     ];
